Adds Go-to-Symbol feature (Ctrl + Shift + O)
